+++ resources/views/livewire/navigation-menu-left.blade.php
@@ -1,9 +1,9 @@
 <aside class="main-sidebar sidebar-dark-primary elevation-4">
     <!-- Brand Logo -->
-    <a href="../../index3.html" class="brand-link">
-        <img src="{{ asset('dist/img/AdminLTELogo.png') }}" alt="AdminLTE Logo" class="brand-image img-circle elevation-3"
-            style="opacity: .8">
-        <span class="brand-text font-weight-light">IMS</span>
+    <a href="{{ route('dashboard') }}" class="brand-link text-center">
+        <img src="{{ asset('dist/img/IMS logo 128 x128.png') }}" alt="AdminLTE Logo" class="img-fluid col-6 text-center "
+            > <br>
+        <span class=" text-wrap  "><small>Inventory Management System</small></span>
     </a>
 
     <!-- Sidebar -->
@@ -14,8 +14,8 @@
                 <img src="{{ asset('assets/images/svgs/icon-user-male.svg') }}" class="img-circle elevation-2" alt="User Image">
             </div>
             <div class="info">
-                <a href="{{ route('profile') }}" class="d-block">{{ Auth::user()->name }}
-                <span class="badge bg-purple" >{{ Auth::user()->role  }}</span>  </a>
+                <a href="{{ route('profile') }}" class="d-block">{{ empty(Auth::user()->name) ? ''  : Auth::user()->name  }}
+                <span class="badge bg-purple" >{{ empty(Auth::user()->role) ? '' : Auth::user()->role  }}</span>  </a>
 
             </div>
         </div>
@@ -28,11 +28,48 @@
                    with font-awesome or any other icon font library -->
 
                 <li class="nav-item">
-                    <a href="{{ route('dashboard') }}" class="nav-link">
+                    <a href="{{ route('dashboard') }}" class="nav-link {{ Route::is('dashboard') ? 'active' : '' }} ">
                         <i class="nav-icon ti ti-dashboard"></i>
                         <p>Dashboard</p>
                     </a>
                 </li>
+                 @if(in_array(Auth::user()->role, ['seller']))
+                <li class="nav-item has-treeview">
+                    <a href="#" class="nav-link">
+                        <i class="nav-icon ti ti-box"></i>
+                        <p>
+                            Inventory
+                            <i class="right fas fa-angle-left"></i>
+                        </p>
+                    </a>
+                    <ul class="nav nav-treeview">
+                        <li class="nav-item">
+                            <a href="{{ route('products') }}" class="nav-link {{ Route::is('products') ? 'active' : '' }}">
+                                <i class="far fa-circle nav-icon"></i>
+                                <p>Products</p>
+                            </a>
+                        </li>
+                        <li class="nav-item">
+                            <a href="{{ route('categories') }}" class="nav-link {{ Route::is('categories') ? 'active' : '' }} ">
+                                <i class="far fa-circle nav-icon"></i>
+                                <p>Categories</p>
+                            </a>
+                        </li>
+                        <li class="nav-item">
+                            <a href="{{ route('suppliers') }}" class="nav-link {{ Route::is('suppliers') ? 'active' : '' }} ">
+                                <i class="far fa-circle nav-icon"></i>
+                                <p>Suppliers</p>
+                            </a>
+                        </li>
+                        <li class="nav-item">
+                            <a href="{{ route('purchases') }}" class="nav-link {{ Route::is('purchases') ? 'active' : '' }} ">
+                                <i class="far fa-circle nav-icon"></i>
+                                <p>Purchases</p>
+                            </a>
+                        </li>
+                    </ul>
+                </li>
+                @endif
 
                 @if(in_array(Auth::user()->role, ['owner', 'system']))
                 <li class="nav-item has-treeview">
@@ -45,25 +82,25 @@
                     </a>
                     <ul class="nav nav-treeview">
                         <li class="nav-item">
-                            <a href="{{ route('products') }}" class="nav-link">
+                            <a href="{{ route('products') }}" class="nav-link {{ Route::is('products') ? 'active' : '' }} ">
                                 <i class="far fa-circle nav-icon"></i>
                                 <p>Products</p>
                             </a>
                         </li>
                         <li class="nav-item">
-                            <a href="{{ route('categories') }}" class="nav-link">
+                            <a href="{{ route('categories') }}" class="nav-link {{ Route::is('categories') ? 'active' : '' }} ">
                                 <i class="far fa-circle nav-icon"></i>
                                 <p>Categories</p>
                             </a>
                         </li>
                         <li class="nav-item">
-                            <a href="{{ route('suppliers') }}" class="nav-link">
+                            <a href="{{ route('suppliers') }}" class="nav-link {{ Route::is('suppliers') ? 'active' : '' }} ">
                                 <i class="far fa-circle nav-icon"></i>
                                 <p>Suppliers</p>
                             </a>
                         </li>
                         <li class="nav-item">
-                            <a href="{{ route('purchases') }}" class="nav-link">
+                            <a href="{{ route('purchases') }}" class="nav-link {{ Route::is('purchases') ? 'active' : '' }} ">
                                 <i class="far fa-circle nav-icon"></i>
                                 <p>Purchases</p>
                             </a>
@@ -83,36 +120,39 @@
                     </a>
                     <ul class="nav nav-treeview">
                         <li class="nav-item">
-                            <a href="{{ route('sales') }}" class="nav-link">
+                            <a href="{{ route('sales') }}" class="nav-link {{ Route::is('sales') ? 'active' : '' }} ">
                                 <i class="far fa-circle nav-icon"></i>
                                 <p>Sales</p>
                             </a>
                         </li>
                         <li class="nav-item">
-                            <a href="{{ route('returns') }}" class="nav-link">
+                            <a href="{{ route('returns') }}" class="nav-link {{ Route::is('returns') ? 'active' : '' }} ">
                                 <i class="far fa-circle nav-icon"></i>
                                 <p>Returns</p>
                             </a>
                         </li>
                         <li class="nav-item">
-                            <a href="{{ route('customers') }}" class="nav-link">
+                            <a href="{{ route('customers') }}" class="nav-link {{ Route::is('customers') ? 'active' : '' }} ">
                                 <i class="far fa-circle nav-icon"></i>
                                 <p>Customers</p>
                             </a>
                         </li>
                         <li class="nav-item">
-                            <a href="{{ route('pos') }}" class="nav-link">
+                            <a href="{{ route('pos') }}" class="nav-link {{ Route::is('pos') ? 'active' : '' }} ">
                                 <i class="far fa-circle nav-icon"></i>
                                 <p>POS</p>
                             </a>
                         </li>
                     </ul>
                 </li>
+
                 @endif
+
+
 
                 @if(in_array(Auth::user()->role, ['owner', 'system']))
                 <li class="nav-item">
-                    <a href="{{ route('expenses') }}" class="nav-link">
+                    <a href="{{ route('expenses') }}" class="nav-link {{ Route::is('expenses') ? 'active' : '' }} ">
                         <i class="nav-icon ti ti-credit-card"></i>
                         <p>Expenses</p>
                     </a>
@@ -130,19 +170,19 @@
                     </a>
                     <ul class="nav nav-treeview">
                         <li class="nav-item">
-                            <a href="{{ route('reports.sales') }}" class="nav-link">
+                            <a href="{{ route('reports.sales') }}" class="nav-link {{ Route::is('reports.sales') ? 'active' : '' }} ">
                                 <i class="far fa-circle nav-icon"></i>
                                 <p>Sales Report</p>
                             </a>
                         </li>
                         <li class="nav-item">
-                            <a href="{{ route('reports.profit') }}" class="nav-link">
+                            <a href="{{ route('reports.profit') }}" class="nav-link {{ Route::is('reports.profit') ? 'active' : '' }} ">
                                 <i class="far fa-circle nav-icon"></i>
                                 <p>Profit Report</p>
                             </a>
                         </li>
                         <li class="nav-item">
-                            <a href="{{ route('reports.stock') }}" class="nav-link">
+                            <a href="{{ route('reports.stock') }}" class="nav-link {{ Route::is('reports.stock') ? 'active' : '' }} ">
                                 <i class="far fa-circle nav-icon"></i>
                                 <p>Stock Report</p>
                             </a>
@@ -162,7 +202,7 @@
                     </a>
                     <ul class="nav nav-treeview">
                         <li class="nav-item">
-                            <a href="{{ route('users') }}" class="nav-link">
+                            <a href="{{ route('users') }}" class="nav-link {{ Route::is('users') ? 'active' : '' }} ">
                                 <i class="far fa-circle nav-icon"></i>
                                 <p>All users</p>
                             </a>
@@ -170,7 +210,7 @@
                     </ul>
                 </li>
                 <li class="nav-item">
-                    <a href="{{ route('audit-logs') }}" class="nav-link">
+                    <a href="{{ route('audit-logs') }}" class="nav-link {{ Route::is('audit-logs') ? 'active' : '' }} ">
                         <i class="nav-icon ti ti-list"></i>
                         <p>
                             Audit Logs